Extract test snippets to fixtures folder
