+++ scripts/fixtures/fragments.go
@@ -1,0 +1,7 @@
+package fixtures
+
+const (
+	StaticIdentifierJavadocCell = `<td class=\"colFirst\"><code>static &lt;E&gt;&nbsp;<a href=\"../../../../com/google/common/collect/ImmutableSet.Builder.html\" title=\"class in com.google.common.collect\">ImmutableSet.Builder</a>&lt;E&gt;</code></td>`
+
+	StaticMethodJavadocRow = `<tr class="altColor"><td class="colFirst"><code>static &lt;E&gt;&nbsp;<a href="../../../../com/google/common/collect/ImmutableSet.Builder.html" title="class in com.google.common.collect">ImmutableSet.Builder</a>&lt;E&gt;</code></td><td class="colLast"><code><strong><a href="../../../../com/google/common/collect/ImmutableSet.html#builder()">builder</a></strong>()</code><div class="block">Returns a new builder.</div></td></tr>`
+)
