+++ scripts/fixtures/full_page.go
@@ -1,0 +1,406 @@
+package fixtures
+
+const MemberClassJavadoc = `
+<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
+<!-- NewPage -->
+<html lang="en">
+<head>
+<!-- Generated by javadoc (1.8.0_251) on Wed Mar 11 23:29:34 PDT 2020 -->
+<title>Member (Java Platform SE 8 )</title>
+<meta name="date" content="2020-03-11">
+<meta name="keywords" content="java.lang.reflect.Member interface">
+<meta name="keywords" content="PUBLIC">
+<meta name="keywords" content="DECLARED">
+<meta name="keywords" content="getDeclaringClass()">
+<meta name="keywords" content="getName()">
+<meta name="keywords" content="getModifiers()">
+<meta name="keywords" content="isSynthetic()">
+<link rel="stylesheet" type="text/css" href="../../../stylesheet.css" title="Style">
+<script type="text/javascript" src="../../../script.js"></script>
+<script>window.ohcglobal || document.write('<script src="/en/dcommon/js/global.js">\x3C/script>')</script></head>
+<body>
+<script type="text/javascript"><!--
+    try {
+        if (location.href.indexOf('is-external=true') == -1) {
+            parent.document.title="Member (Java Platform SE 8 )";
+        }
+    }
+    catch(err) {
+    }
+//-->
+var methods = {"i0":6,"i1":6,"i2":6,"i3":6};
+var tabs = {65535:["t0","All Methods"],2:["t2","Instance Methods"],4:["t3","Abstract Methods"]};
+var altColor = "altColor";
+var rowColor = "rowColor";
+var tableTab = "tableTab";
+var activeTableTab = "activeTableTab";
+</script>
+<noscript>
+<div>JavaScript is disabled on your browser.</div>
+</noscript>
+<!-- ========= START OF TOP NAVBAR ======= -->
+<div class="topNav"><a name="navbar.top">
+<!--   -->
+</a>
+<div class="skipNav"><a href="#skip.navbar.top" title="Skip navigation links">Skip navigation links</a></div>
+<a name="navbar.top.firstrow">
+<!--   -->
+</a>
+<ul class="navList" title="Navigation">
+<li><a href="../../../overview-summary.html">Overview</a></li>
+<li><a href="package-summary.html">Package</a></li>
+<li class="navBarCell1Rev">Class</li>
+<li><a href="class-use/Member.html">Use</a></li>
+<li><a href="package-tree.html">Tree</a></li>
+<li><a href="../../../deprecated-list.html">Deprecated</a></li>
+<li><a href="../../../index-files/index-1.html">Index</a></li>
+<li><a href="../../../help-doc.html">Help</a></li>
+</ul>
+<div class="aboutLanguage"><strong>Java&trade;&nbsp;Platform<br>Standard&nbsp;Ed.&nbsp;8</strong></div>
+</div>
+<div class="subNav">
+<ul class="navList">
+<li><a href="../../../java/lang/reflect/MalformedParametersException.html" title="class in java.lang.reflect"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
+<li><a href="../../../java/lang/reflect/Method.html" title="class in java.lang.reflect"><span class="typeNameLink">Next&nbsp;Class</span></a></li>
+</ul>
+<ul class="navList">
+<li><a href="../../../index.html?java/lang/reflect/Member.html" target="_top">Frames</a></li>
+<li><a href="Member.html" target="_top">No&nbsp;Frames</a></li>
+</ul>
+<ul class="navList" id="allclasses_navbar_top">
+<li><a href="../../../allclasses-noframe.html">All&nbsp;Classes</a></li>
+</ul>
+<div>
+<script type="text/javascript"><!--
+  allClassesLink = document.getElementById("allclasses_navbar_top");
+  if(window==top) {
+    allClassesLink.style.display = "block";
+  }
+  else {
+    allClassesLink.style.display = "none";
+  }
+  //-->
+</script>
+</div>
+<div>
+<ul class="subNavList">
+<li>Summary:&nbsp;</li>
+<li>Nested&nbsp;|&nbsp;</li>
+<li><a href="#field.summary">Field</a>&nbsp;|&nbsp;</li>
+<li>Constr&nbsp;|&nbsp;</li>
+<li><a href="#method.summary">Method</a></li>
+</ul>
+<ul class="subNavList">
+<li>Detail:&nbsp;</li>
+<li><a href="#field.detail">Field</a>&nbsp;|&nbsp;</li>
+<li>Constr&nbsp;|&nbsp;</li>
+<li><a href="#method.detail">Method</a></li>
+</ul>
+</div>
+<a name="skip.navbar.top">
+<!--   -->
+</a></div>
+<!-- ========= END OF TOP NAVBAR ========= -->
+<!-- ======== START OF CLASS DATA ======== -->
+<div class="header">
+<div class="subTitle">compact1, compact2, compact3</div>
+<div class="subTitle">java.lang.reflect</div>
+<h2 title="Interface Member" class="title">Interface Member</h2>
+</div>
+<div class="contentContainer">
+<div class="description">
+<ul class="blockList">
+<li class="blockList">
+<dl>
+<dt>All Known Implementing Classes:</dt>
+<dd><a href="../../../java/lang/reflect/Constructor.html" title="class in java.lang.reflect">Constructor</a>, <a href="../../../java/lang/reflect/Executable.html" title="class in java.lang.reflect">Executable</a>, <a href="../../../java/lang/reflect/Field.html" title="class in java.lang.reflect">Field</a>, <a href="../../../java/lang/reflect/Method.html" title="class in java.lang.reflect">Method</a></dd>
+</dl>
+<hr>
+<br>
+<pre>public interface <span class="typeNameLabel">Member</span></pre>
+<div class="block">Member is an interface that reflects identifying information about
+ a single member (a field or a method) or a constructor.</div>
+<dl>
+<dt><span class="seeLabel">See Also:</span></dt>
+<dd><a href="../../../java/lang/Class.html" title="class in java.lang"><code>Class</code></a>,
+<a href="../../../java/lang/reflect/Field.html" title="class in java.lang.reflect"><code>Field</code></a>,
+<a href="../../../java/lang/reflect/Method.html" title="class in java.lang.reflect"><code>Method</code></a>,
+<a href="../../../java/lang/reflect/Constructor.html" title="class in java.lang.reflect"><code>Constructor</code></a></dd>
+</dl>
+</li>
+</ul>
+</div>
+<div class="summary">
+<ul class="blockList">
+<li class="blockList">
+<!-- =========== FIELD SUMMARY =========== -->
+<ul class="blockList">
+<li class="blockList"><a name="field.summary">
+<!--   -->
+</a>
+<h3>Field Summary</h3>
+<table class="memberSummary" border="0" cellpadding="3" cellspacing="0" summary="Field Summary table, listing fields, and an explanation">
+<caption><span>Fields</span><span class="tabEnd">&nbsp;</span></caption>
+<tr>
+<th class="colFirst" scope="col">Modifier and Type</th>
+<th class="colLast" scope="col">Field and Description</th>
+</tr>
+<tr class="altColor">
+<td class="colFirst"><code>static int</code></td>
+<td class="colLast"><code><span class="memberNameLink"><a href="../../../java/lang/reflect/Member.html#DECLARED">DECLARED</a></span></code>
+<div class="block">Identifies the set of declared members of a class or interface.</div>
+</td>
+</tr>
+<tr class="rowColor">
+<td class="colFirst"><code>static int</code></td>
+<td class="colLast"><code><span class="memberNameLink"><a href="../../../java/lang/reflect/Member.html#PUBLIC">PUBLIC</a></span></code>
+<div class="block">Identifies the set of all public members of a class or interface,
+ including inherited members.</div>
+</td>
+</tr>
+</table>
+</li>
+</ul>
+<!-- ========== METHOD SUMMARY =========== -->
+<ul class="blockList">
+<li class="blockList"><a name="method.summary">
+<!--   -->
+</a>
+<h3>Method Summary</h3>
+<table class="memberSummary" border="0" cellpadding="3" cellspacing="0" summary="Method Summary table, listing methods, and an explanation">
+<caption><span id="t0" class="activeTableTab"><span>All Methods</span><span class="tabEnd">&nbsp;</span></span><span id="t2" class="tableTab"><span><a href="javascript:show(2);">Instance Methods</a></span><span class="tabEnd">&nbsp;</span></span><span id="t3" class="tableTab"><span><a href="javascript:show(4);">Abstract Methods</a></span><span class="tabEnd">&nbsp;</span></span></caption>
+<tr>
+<th class="colFirst" scope="col">Modifier and Type</th>
+<th class="colLast" scope="col">Method and Description</th>
+</tr>
+<tr id="i0" class="altColor">
+<td class="colFirst"><code><a href="../../../java/lang/Class.html" title="class in java.lang">Class</a>&lt;?&gt;</code></td>
+<td class="colLast"><code><span class="memberNameLink"><a href="../../../java/lang/reflect/Member.html#getDeclaringClass--">getDeclaringClass</a></span>()</code>
+<div class="block">Returns the Class object representing the class or interface
+ that declares the member or constructor represented by this Member.</div>
+</td>
+</tr>
+<tr id="i1" class="rowColor">
+<td class="colFirst"><code>int</code></td>
+<td class="colLast"><code><span class="memberNameLink"><a href="../../../java/lang/reflect/Member.html#getModifiers--">getModifiers</a></span>()</code>
+<div class="block">Returns the Java language modifiers for the member or
+ constructor represented by this Member, as an integer.</div>
+</td>
+</tr>
+<tr id="i2" class="altColor">
+<td class="colFirst"><code><a href="../../../java/lang/String.html" title="class in java.lang">String</a></code></td>
+<td class="colLast"><code><span class="memberNameLink"><a href="../../../java/lang/reflect/Member.html#getName--">getName</a></span>()</code>
+<div class="block">Returns the simple name of the underlying member or constructor
+ represented by this Member.</div>
+</td>
+</tr>
+<tr id="i3" class="rowColor">
+<td class="colFirst"><code>boolean</code></td>
+<td class="colLast"><code><span class="memberNameLink"><a href="../../../java/lang/reflect/Member.html#isSynthetic--">isSynthetic</a></span>()</code>
+<div class="block">Returns <code>true</code> if this member was introduced by
+ the compiler; returns <code>false</code> otherwise.</div>
+</td>
+</tr>
+</table>
+</li>
+</ul>
+</li>
+</ul>
+</div>
+<div class="details">
+<ul class="blockList">
+<li class="blockList">
+<!-- ============ FIELD DETAIL =========== -->
+<ul class="blockList">
+<li class="blockList"><a name="field.detail">
+<!--   -->
+</a>
+<h3>Field Detail</h3>
+<a name="PUBLIC">
+<!--   -->
+</a>
+<ul class="blockList">
+<li class="blockList">
+<h4>PUBLIC</h4>
+<pre>static final&nbsp;int PUBLIC</pre>
+<div class="block">Identifies the set of all public members of a class or interface,
+ including inherited members.</div>
+<dl>
+<dt><span class="seeLabel">See Also:</span></dt>
+<dd><a href="../../../constant-values.html#java.lang.reflect.Member.PUBLIC">Constant Field Values</a></dd>
+</dl>
+</li>
+</ul>
+<a name="DECLARED">
+<!--   -->
+</a>
+<ul class="blockListLast">
+<li class="blockList">
+<h4>DECLARED</h4>
+<pre>static final&nbsp;int DECLARED</pre>
+<div class="block">Identifies the set of declared members of a class or interface.
+ Inherited members are not included.</div>
+<dl>
+<dt><span class="seeLabel">See Also:</span></dt>
+<dd><a href="../../../constant-values.html#java.lang.reflect.Member.DECLARED">Constant Field Values</a></dd>
+</dl>
+</li>
+</ul>
+</li>
+</ul>
+<!-- ============ METHOD DETAIL ========== -->
+<ul class="blockList">
+<li class="blockList"><a name="method.detail">
+<!--   -->
+</a>
+<h3>Method Detail</h3>
+<a name="getDeclaringClass--">
+<!--   -->
+</a>
+<ul class="blockList">
+<li class="blockList">
+<h4>getDeclaringClass</h4>
+<pre><a href="../../../java/lang/Class.html" title="class in java.lang">Class</a>&lt;?&gt;&nbsp;getDeclaringClass()</pre>
+<div class="block">Returns the Class object representing the class or interface
+ that declares the member or constructor represented by this Member.</div>
+<dl>
+<dt><span class="returnLabel">Returns:</span></dt>
+<dd>an object representing the declaring class of the
+ underlying member</dd>
+</dl>
+</li>
+</ul>
+<a name="getName--">
+<!--   -->
+</a>
+<ul class="blockList">
+<li class="blockList">
+<h4>getName</h4>
+<pre><a href="../../../java/lang/String.html" title="class in java.lang">String</a>&nbsp;getName()</pre>
+<div class="block">Returns the simple name of the underlying member or constructor
+ represented by this Member.</div>
+<dl>
+<dt><span class="returnLabel">Returns:</span></dt>
+<dd>the simple name of the underlying member</dd>
+</dl>
+</li>
+</ul>
+<a name="getModifiers--">
+<!--   -->
+</a>
+<ul class="blockList">
+<li class="blockList">
+<h4>getModifiers</h4>
+<pre>int&nbsp;getModifiers()</pre>
+<div class="block">Returns the Java language modifiers for the member or
+ constructor represented by this Member, as an integer.  The
+ Modifier class should be used to decode the modifiers in
+ the integer.</div>
+<dl>
+<dt><span class="returnLabel">Returns:</span></dt>
+<dd>the Java language modifiers for the underlying member</dd>
+<dt><span class="seeLabel">See Also:</span></dt>
+<dd><a href="../../../java/lang/reflect/Modifier.html" title="class in java.lang.reflect"><code>Modifier</code></a></dd>
+</dl>
+</li>
+</ul>
+<a name="isSynthetic--">
+<!--   -->
+</a>
+<ul class="blockListLast">
+<li class="blockList">
+<h4>isSynthetic</h4>
+<pre>boolean&nbsp;isSynthetic()</pre>
+<div class="block">Returns <code>true</code> if this member was introduced by
+ the compiler; returns <code>false</code> otherwise.</div>
+<dl>
+<dt><span class="returnLabel">Returns:</span></dt>
+<dd>true if and only if this member was introduced by
+ the compiler.</dd>
+<dt><span class="simpleTagLabel">Since:</span></dt>
+<dd>1.5</dd>
+<dt><span class="simpleTagLabel">See <cite> The Java&trade; Language Specification</cite>:</span></dt>
+<dd>13.1 The Form of a Binary</dd>
+</dl>
+</li>
+</ul>
+</li>
+</ul>
+</li>
+</ul>
+</div>
+</div>
+<!-- ========= END OF CLASS DATA ========= -->
+<!-- ======= START OF BOTTOM NAVBAR ====== -->
+<div class="bottomNav"><a name="navbar.bottom">
+<!--   -->
+</a>
+<div class="skipNav"><a href="#skip.navbar.bottom" title="Skip navigation links">Skip navigation links</a></div>
+<a name="navbar.bottom.firstrow">
+<!--   -->
+</a>
+<ul class="navList" title="Navigation">
+<li><a href="../../../overview-summary.html">Overview</a></li>
+<li><a href="package-summary.html">Package</a></li>
+<li class="navBarCell1Rev">Class</li>
+<li><a href="class-use/Member.html">Use</a></li>
+<li><a href="package-tree.html">Tree</a></li>
+<li><a href="../../../deprecated-list.html">Deprecated</a></li>
+<li><a href="../../../index-files/index-1.html">Index</a></li>
+<li><a href="../../../help-doc.html">Help</a></li>
+</ul>
+<div class="aboutLanguage"><strong>Java&trade;&nbsp;Platform<br>Standard&nbsp;Ed.&nbsp;8</strong></div>
+</div>
+<div class="subNav">
+<ul class="navList">
+<li><a href="../../../java/lang/reflect/MalformedParametersException.html" title="class in java.lang.reflect"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
+<li><a href="../../../java/lang/reflect/Method.html" title="class in java.lang.reflect"><span class="typeNameLink">Next&nbsp;Class</span></a></li>
+</ul>
+<ul class="navList">
+<li><a href="../../../index.html?java/lang/reflect/Member.html" target="_top">Frames</a></li>
+<li><a href="Member.html" target="_top">No&nbsp;Frames</a></li>
+</ul>
+<ul class="navList" id="allclasses_navbar_bottom">
+<li><a href="../../../allclasses-noframe.html">All&nbsp;Classes</a></li>
+</ul>
+<div>
+<script type="text/javascript"><!--
+  allClassesLink = document.getElementById("allclasses_navbar_bottom");
+  if(window==top) {
+    allClassesLink.style.display = "block";
+  }
+  else {
+    allClassesLink.style.display = "none";
+  }
+  //-->
+</script>
+</div>
+<div>
+<ul class="subNavList">
+<li>Summary:&nbsp;</li>
+<li>Nested&nbsp;|&nbsp;</li>
+<li><a href="#field.summary">Field</a>&nbsp;|&nbsp;</li>
+<li>Constr&nbsp;|&nbsp;</li>
+<li><a href="#method.summary">Method</a></li>
+</ul>
+<ul class="subNavList">
+<li>Detail:&nbsp;</li>
+<li><a href="#field.detail">Field</a>&nbsp;|&nbsp;</li>
+<li>Constr&nbsp;|&nbsp;</li>
+<li><a href="#method.detail">Method</a></li>
+</ul>
+</div>
+<a name="skip.navbar.bottom">
+<!--   -->
+</a></div>
+<!-- ======== END OF BOTTOM NAVBAR ======= -->
+<p class="legalCopy"><small><font size="-1"> <a href="http://bugreport.sun.com/bugreport/">Submit a bug or feature</a> <br>For further API reference and developer documentation, see <a href="https://docs.oracle.com/javase/8/docs/index.html" target="_blank">Java SE Documentation</a>. That documentation contains more detailed, developer-targeted descriptions, with conceptual overviews, definitions of terms, workarounds, and working code examples.<br> <a href="../../../../legal/cpyr.html">Copyright</a> &#x00a9; 1993, 2020, Oracle and/or its affiliates.  All rights reserved. Use is subject to <a href="http://download.oracle.com/otndocs/jcp/java_se-8-mrel-spec/license.html">license terms</a>. Also see the <a target="_blank" href="http://www.oracle.com/technetwork/java/redist-137594.html">documentation redistribution policy</a>. </font></small></p>
+<!-- Start SiteCatalyst code   -->
+<script type="application/javascript" src="https://www.oracleimg.com/us/assets/metrics/ora_docs.js"></script>
+<!-- End SiteCatalyst code -->
+<noscript>
+<p>Scripting on this page tracks web page traffic, but does not change the content in any way.</p>
+</noscript>
+</body>
+</html>
+`
